projects/knowledge_base_agent/requirements.txt
chromadb
langchain
langchain-openai
langchain-community
langchain-huggingface
pypdf
tiktoken
python-dotenv
openai
sentence-transformers